Make the script POSIX compliant
The script has the shebang '#!/bin/sh' which can be a symlink to any
POSIX-compliant shell, however the script made use of a variety of
bash-specific features. The script is now fully POSIX-compliant besides
the obvious (the use of the stow(1) utility), and has been successfully
tested on the dash shell.